+++ src/palau/lims/impress/reportview.py
@@ -4,8 +4,10 @@
 #
 # Copyright 2023 Beyond Essential Systems Pty Ltd
 
-import json
+import copy
 
+import collections
+import json
 from bika.lims import api
 from bika.lims.api import mail
 from bika.lims.utils import get_link
@@ -22,7 +24,6 @@ from senaite.impress.decorators import returns_super_model
 from senaite.patient import api as patient_api
 from senaite.patient.config import SEXES
 from weasyprint.compat import base64_encode
-
 
 BOOL_TEXTS = {True: _("Yes"), False: _("No")}
 
@@ -114,10 +115,9 @@ class DefaultReportView(SingleReportView):
         """Returns the Department title
         """
         sample = api.get_object(sample)
-        department = sample.getField("WardDepartment").get(sample)
+        department = sample.getWardDepartment()
         if not department:
             return ""
-
         return api.get_title(department)
 
     def get_analyses(self, model_or_collection, parts=False):
@@ -197,26 +197,6 @@ class DefaultReportView(SingleReportView):
         """
         analyses = self.get_analyses(model_or_collection, parts=parts)
         return self.group_items_by("Category", analyses)
-
-    def is_valid_status(self, analysis):
-        """Returns whether the analysis object or brain is in a valid status
-        """
-        invalid = ["retracted", "rejected", "cancelled"]
-        return api.get_review_status(analysis) not in invalid
-
-    def is_results_in_progress(self, sample):
-        """Returns true if there are analyses not yet verified
-        """
-        def in_progress(analysis):
-            return not analysis.getDateVerified()
-
-        # Exclude invalid analyses
-        analyses = self.get_analyses(sample)
-        analyses = filter(self.is_valid_status, analyses)
-
-        # Ensure all valid analyses are in progress
-        analyses = map(in_progress, analyses)
-        return any(analyses)
 
     def get_formatted_result(self, sample_model, analysis):
         """Returns the result of the analysis properly formatted
@@ -484,18 +464,30 @@ class DefaultReportView(SingleReportView):
         return filter(None, submitters)
 
     def get_results_interpretations(self, model):
-        """Mimics the function analysisrequest.model.get_resultsinterpretation
-        from senaite.impress, but injects the keys "user" and "fullname"
+        """Returns the result interpretations
         """
-        ri_by_depts = model.ResultsInterpretationDepts
+        # do a hard copy to prevent persistent changes
+        interpretations = copy.deepcopy(model.getResultsInterpretationDepts())
+
+        # group by user
+        groups = collections.OrderedDict()
+        for interpretation in interpretations:
+            user = interpretation.get("user", "")
+            groups.setdefault(user, []).append(interpretation)
+
         out = []
-        for ri in ri_by_depts:
-            dept = ri.get("uid", "")
-            user = ri.get("user") or ""
+        for user, items in groups.items():
+            # get the comments from this user
+            comments = [item.get("richtext", "").strip() for item in items]
+            # bail out those with no value or empty
+            comments = filter(None, comments)
+            if not comments:
+                continue
+
             out.append({
-                "title": getattr(dept, "title", ""),
-                "richtext": ri.get("richtext", ""),
                 "user": user,
                 "fullname": get_fullname(user),
+                "richtext": "".join(comments),
             })
+
         return out